Six conflicts, all from upstream refactors landing on code this branch had
extended for floating point.

lib/Printer/SMTLIB2Printer.cpp: #721 folded the version1 and version2
traversals into one SMTLIB_Print1() in SMTLIBPrinter.cpp, while this branch
had grown the version2 traversal the whole FP switch. Moved those cases --
the rounded operations, min/max, the to_fp family, fp.to_ubv/to_sbv, and the
BVCONST branch that prints a float's packed bits as (fp ...) and a rounding
mode by name -- into the shared traversal, along with the
printRoundingModeSMTLIB2() helper they need. They are not conditioned on the
dialect flag: SMT-LIB1 has no FP theory, so those kinds only ever arrive with
it clear. outputFloatingPointSMTLIB2() and roundingModeName() stay put;
CounterExample and FPPrintBack_Test call them directly.

lib/Interface/c_interface.cpp: #708 deduplicated the entry points. The
mgr()/wrap() helpers it adds and this branch's source-sort helpers were an
additive conflict, so both are kept. vc_andExpr, vc_orExpr, vc_xorExpr,
vc_impliesExpr and vc_iffExpr now delegate to createBinaryNode() as upstream
wrote them, with AND/OR/XOR/IMPLIES/IFF added to the arm that already routed
NAND/NOR through requireBooleanOperand() -- so the operand check this branch
added survives the dedup, and its message becomes the same "Boolean
connective" wording nand/nor already used. vc_varExpr keeps this branch's
source-sort body and takes upstream's mgr(vc).

include/stp/cpp_interface.h: #710 removed isBitVectorFunction() and
isBooleanFunction(), which were declared but never defined or called. Took
the deletion, keeping this branch's comment on functionReturnType() because
it points at functionReturnSourceSort(), which upstream does not have.

lib/Sat/CryptoMinisat5.cpp: upstream replaced the unused simplify() result
in getFixedCountWithAssumptions() with set_no_bve(); this branch had only
marked that variable [[maybe_unused]]. Upstream's version supersedes it --
the variable is gone entirely.

tools/measure_constantbitprop and tools/test_constantbitprop: #716 deleted
both tools, replaced by the test suite and propagator_bench. This branch had
only added [[maybe_unused]] annotations to them, so the deletion stands.

Verified: build clean with no warnings in STP's own sources, ctest 90/90,
query-files 269/269 (268 before the merge; #719 adds one).

+++ include/stp/AST/ASTInternal.h
@@ -184,7 +184,7 @@ public:
     }
   }
 
-  unsigned GetNodeNum() const { return node_uid; }
+  uint64_t GetNodeNum() const { return node_uid; }
 
   virtual bool isSimplified() const { return false; }
