+++ lib/AST/ASTmisc.cpp
@@ -92,7 +92,7 @@ bool arithless(const ASTNode& n1, const ASTNode& n2)
 }
 
 // counts the number of reads. Shortcut when we get to the limit.
-void numberOfReadsLessThan(const ASTNode& n, std::unordered_set<int>& visited,
+void numberOfReadsLessThan(const ASTNode& n, std::unordered_set<uint64_t>& visited,
                            int& soFar, const int limit)
 {
   if (n.isAtom())
@@ -116,7 +116,7 @@ void numberOfReadsLessThan(const ASTNode& n, std::unordered_set<int>& visited,
 // True if the number of reads in "n" is less than "limit"
 bool numberOfReadsLessThan(const ASTNode& n, int limit)
 {
-  std::unordered_set<int> visited;
+  std::unordered_set<uint64_t> visited;
   int reads = 0;
   numberOfReadsLessThan(n, visited, reads, limit);
   return reads < limit;
